+++ views/global/index.php
@@ -12,11 +12,20 @@ use humhub\modules\tasks\models\forms\TaskFilter;
 ?>
 
 <div class="container">
-    <div class="panel panel-default task-overview">
-        <?= TaskFilterNavigation::widget(['options' => ['style' => 'border-radius:4px;background-color:'.$this->theme->variable('background-color-secondary')]]) ?>
+    <div class="panel panel-default">
+        <div class="panel-heading">
+            <i class="fa fa-tasks"></i> <?= Yii::t('TasksModule.base', '<strong>Task</strong> Overview'); ?>
+        </div>
 
-        <div id="filter-tasks-list" class="panel-body">
-            <?= TaskSearchList::widget(['filter' => new TaskFilter()]) ?>
+
+
+        <div class="task-overview">
+
+            <?= TaskFilterNavigation::widget(['options' => ['style' => 'border-radius:4px;background-color:' . $this->theme->variable('background-color-secondary')]]) ?>
+
+            <div id="filter-tasks-list" class="panel-body">
+                <?= TaskSearchList::widget(['filter' => new TaskFilter()]) ?>
+            </div>
         </div>
     </div>
 </div>