+++ widgets/search/TaskSearchList.php
@@ -17,6 +17,7 @@ namespace humhub\modules\tasks\widgets\search;
 use humhub\components\Widget;
 use humhub\modules\tasks\models\forms\TaskFilter;
 use humhub\modules\tasks\models\Task;
+use humhub\modules\tasks\Module;
 use Yii;
 use yii\data\ActiveDataProvider;
 use yii\widgets\ListView;
@@ -32,10 +33,13 @@ class TaskSearchList extends Widget
 
     public function run()
     {
+        /* @var $module Module */
+        $module = Yii::$app->getModule('tasks');
+
         $tasksProvider = new ActiveDataProvider([
             'query' => $this->filter->query(),
             'pagination' => [
-                'pageSize' => 30,
+                'pageSize' => $module->searchPaginationSize,
                 'route' => '/tasks/global/filter'
             ],
         ]);
@@ -47,6 +51,7 @@ class TaskSearchList extends Widget
                 'contentContainer' => $this->filter->contentContainer,
                 'canEdit' => $this->canEdit
             ],
+            'summary'=>'',
             'options' => [
                 'tag' => 'ul',
                 'class' => 'media-list'