+++ widgets/search/views/taskFilterNavigation.php
@@ -19,7 +19,7 @@ $title = $panels[TaskFilterNavigation::PANEL_POSITION_TOP][0]['filters'];
 $checkboxes = $panels[TaskFilterNavigation::PANEL_POSITION_BOTTOM_LEFT];
 $picker = $panels[TaskFilterNavigation::PANEL_POSITION_BOTTOM_CENTER];
 
-
+$titleFilter = $title[0];
 ?>
 
 <?= Html::beginTag('div', $options) ?>
@@ -31,12 +31,15 @@ $picker = $panels[TaskFilterNavigation::PANEL_POSITION_BOTTOM_CENTER];
 
     </div>
 
-<div class="filter-root">
+<div class="filter-root clearfix" style="position:relative;padding-top:0;padding-bottom:0">
     <div class="row task-bottom-panel">
         <?= FilterPanel::widget(['blocks' => $checkboxes, 'span' => 2]) ?>
         <?= FilterPanel::widget(['blocks' => $picker, 'span' => 2]) ?>
+
     </div>
+    <div id="task-search-loader" style="position:absolute;bottom:0;right:0"></div>
 </div>
+
 
 
 <?= Html::endTag('div') ?>