+++ widgets/views/myTasks.php
@@ -18,23 +18,13 @@ use humhub\modules\tasks\helpers\TaskUrl;
 
     <div class="panel-heading">
         <i class="fa fa-tasks"></i> <?= Yii::t('TasksModule.widgets_views_myTasks', '<strong>Your</strong> tasks'); ?>
+        <small><a style="font-size:0.9em;color:<?= $this->theme->variable('info'); ?>" href="<?=  TaskUrl::globalView() ?>">(<?= Yii::t('TasksModule.base', 'view all'); ?>)</a></small>
         <?= PanelMenu::widget(['id' => 'task-my-tasks-snippet']); ?>
     </div>
 
     <div class="panel-body" style="padding:0px;">
         <hr style="margin:0px">
         <ul class="media-list">
-            <a href="<?= TaskUrl::globalView() ?>">
-                <li style="background-color:<?= Yii::$app->view->theme->variable('background-color-secondary')?>;border-left:0">
-                    <div class="media">
-                        <div class="media-body text-break">
-                            <strong>
-                                <i class="fa fa-globe"></i>     <?= Yii::t('TasksModule.base', 'View all Tasks') ?>
-                            </strong>
-                        </div>
-                    </div>
-                </li>
-            </a>
             <?php foreach ($taskEntries as $entry) : ?>
                 <?php $color = $entry->getColor() ? $entry->getColor() : $this->theme->variable('info') ?>
                 <a href="<?= $entry->getUrl() ?>">